requirements.txt
PyQt5>=5.15.0
openslide-python>=3.4.1
Pillow>=8.0.0
pyinstaller>=5.0.0
setuptools>=45.0.0
wheel>=0.37.0
numpy>=1.20.0 